refactor :config 결합 및 Interceptor범위 수정
