--- backend/src/main/java/org/donggle/backend/config/WebConfiguration.java
@@ -1,43 +1,0 @@
-package org.donggle.backend.config;
-
-import lombok.RequiredArgsConstructor;
-import org.donggle.backend.application.repository.TokenRepository;
-import org.donggle.backend.auth.JwtTokenProvider;
-import org.donggle.backend.auth.presentation.AuthInterceptor;
-import org.donggle.backend.auth.presentation.RefreshTokenAuthInterceptor;
-import org.donggle.backend.auth.presentation.TokenResolver;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.web.method.support.HandlerMethodArgumentResolver;
-import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
-import org.springframework.web.servlet.config.annotation.ResourceHandlerRegistry;
-import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
-
-import java.util.List;
-
-@RequiredArgsConstructor
-@Configuration
-public class WebConfiguration implements WebMvcConfigurer {
-
-    private final JwtTokenProvider jwtTokenProvider;
-    private final TokenRepository tokenRepository;
-
-    @Override
-    public void addInterceptors(final InterceptorRegistry registry) {
-        registry.addInterceptor(new AuthInterceptor(jwtTokenProvider))
-                .addPathPatterns("/api/**");
-
-        registry.addInterceptor(new RefreshTokenAuthInterceptor(jwtTokenProvider, tokenRepository))
-                .addPathPatterns("/api/auth/accessToken");
-    }
-
-    @Override
-    public void addResourceHandlers(final ResourceHandlerRegistry registry) {
-        registry.addResourceHandler("/*")
-                .addResourceLocations("classpath:/static/");
-    }
-
-    @Override
-    public void addArgumentResolvers(final List<HandlerMethodArgumentResolver> resolvers) {
-        resolvers.add(new TokenResolver(jwtTokenProvider));
-    }
-}
