+++ backend/src/main/java/org/donggle/backend/ui/config/WebMvcConfig.java
@@ -1,12 +1,27 @@
 package org.donggle.backend.ui.config;
 
+import lombok.RequiredArgsConstructor;
+import org.donggle.backend.application.repository.TokenRepository;
+import org.donggle.backend.auth.JwtTokenProvider;
+import org.donggle.backend.auth.presentation.AuthInterceptor;
+import org.donggle.backend.auth.presentation.RefreshTokenAuthInterceptor;
+import org.donggle.backend.auth.presentation.TokenArgumentResolver;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.http.HttpHeaders;
+import org.springframework.web.method.support.HandlerMethodArgumentResolver;
 import org.springframework.web.servlet.config.annotation.CorsRegistry;
+import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
 import org.springframework.web.servlet.config.annotation.WebMvcConfigurer;
 
+import java.util.List;
+
 @Configuration
+@RequiredArgsConstructor
 public class WebMvcConfig implements WebMvcConfigurer {
+
+    private final JwtTokenProvider jwtTokenProvider;
+    private final TokenRepository tokenRepository;
+
     @Override
     public void addCorsMappings(final CorsRegistry registry) {
         registry.addMapping("/**")
@@ -16,4 +31,19 @@ public class WebMvcConfig implements WebMvcConfigurer {
                 .exposedHeaders(HttpHeaders.LOCATION)
                 .maxAge(3600);
     }
+
+    @Override
+    public void addInterceptors(final InterceptorRegistry registry) {
+        registry.addInterceptor(new AuthInterceptor(jwtTokenProvider))
+                .addPathPatterns("/*");
+
+        registry.addInterceptor(new RefreshTokenAuthInterceptor(jwtTokenProvider, tokenRepository))
+                .addPathPatterns("/token/refresh");
+    }
+
+    @Override
+    public void addArgumentResolvers(final List<HandlerMethodArgumentResolver> resolvers) {
+        resolvers.add(new TokenArgumentResolver(jwtTokenProvider));
+    }
 }
+
